apps/km-docs-lambda/requirements.txt
boto3
beautifulsoup4
dicttoxml
google-api-python-client
mkdocs-video==1.3.0
certifi>=2023.7.22 # not directly required, pinned by Snyk to avoid a vulnerability
protobuf>=3.18.3 # not directly required, pinned by Snyk to avoid a vulnerability
requests>=2.31.0 # not directly required, pinned by Snyk to avoid a vulnerability
rsa>=4.7 # not directly required, pinned by Snyk to avoid a vulnerability
setuptools>=65.5.1 # not directly required, pinned by Snyk to avoid a vulnerability
